rename module
(cherry picked from commit c253adac59242979978eb4397af45928b67a16d3)
